+++ packages/extension-polkagate/src/fullscreen/manageProxies/consts.ts
@@ -4,7 +4,7 @@
 import type React from 'react';
 import type { ProxyTypes } from '@polkadot/extension-polkagate/util/types';
 
-import { Chart2, CloseCircle, Data2, HierarchySquare2, Judge, People, Profile2User, Setting2, ShieldTick, Slash } from 'iconsax-react';
+import { Chart2, CloseCircle, Data2, HierarchySquare2, Judge, People, Profile2User, Setting2, ShieldTick, Slash, UserSquare, Wallet2 } from 'iconsax-react';
 
 import { TRANSACTION_FLOW_STEPS } from '@polkadot/extension-polkagate/src/util/constants';
 
@@ -18,8 +18,12 @@ export const STEPS = {
 
 export const PROXY_ICONS: Record<ProxyTypes, React.ElementType> = {
   Any: Data2,
+  AssetManager: Setting2,
+  AssetOwner: Profile2User,
+  Assets: Wallet2,
   Auction: HierarchySquare2,
   CancelProxy: CloseCircle,
+  Collator: UserSquare,
   Governance: Judge,
   IdentityJudgement: ShieldTick,
   NominationPools: People,